refactor: remove PaymentDetails, PaymentUpload, and ShippingInfo components
