+++ apps/web/src/components/cart/CartItemImage.tsx
@@ -2,34 +2,21 @@
 
 import Image from "next/image";
 import React from "react";
-import { useQuery } from "@tanstack/react-query";
 
 interface Props {
+  imageUrl?: string;
   productId: number;
   alt: string;
 }
 
-export default function CartItemImage({ productId, alt }: Props) {
-  // Get product details to get the correct image
-  const { data: productsData } = useQuery({
-    queryKey: ["products"],
-    queryFn: () =>
-      import("@/services/products.service").then((s) =>
-        s.productsService.getProducts()
-      ),
-  });
-
-  // Find the product with matching ID
-  const product = productsData?.products?.find(
-    (p) => parseInt(p.id) === productId
-  );
-  const imageUrl =
-    product?.imageUrl || `https://picsum.photos/seed/${productId}/200/200`;
+export default function CartItemImage({ imageUrl, productId, alt }: Props) {
+  // Use provided imageUrl or fallback to placeholder
+  const imgSrc = imageUrl || `https://picsum.photos/seed/${productId}/200/200`;
 
   return (
     <div className="w-20 h-20 md:w-24 md:h-24 bg-white rounded-xl flex-shrink-0 overflow-hidden relative shadow-sm">
       <Image
-        src={imageUrl}
+        src={imgSrc}
         alt={alt}
         fill
         sizes="(max-width: 640px) 80px, 96px"
